modules/job-manager: create header for module context
